Fix rendering bug on the grid
In this commit, I addressed what is probably the most challenging issue
of making an etch-a-sketch game -- rendering the correct number of
tiles/cells on the provided space. On previous attempts, I tried to let
grid handle all the cell positioning, but this was producing results
that were all either on a row or column only. I was trying to make the
rendering O(N), but this was causing that issue. The solution I found to
this issue was to use an O(N^2) algorithm to populate each row at a
time. It takes longer, but as we are not generating millions of items,
this should not be too task intensive.

The other issue I was running into was that even though the algorithm
was now rendering the correct number of cells on each row, each cell was
not rendering with the correct dimensions. A closer inspection at the
debugger was throwing out correct values for the sizes. However, when
hovering over each cell, they would have an overall dimension that was
the correct one, but each was displayed with the wrong dimensions. I
solved this issue by changing both the grid container and grid tiles to
use the alternate box model. This did two things:

1. It ensured that each cell was rendered with the right dimensions both
   visually and container-wise.
2. It ensured the container never exceeded the max-width and max-height
   specified on it.

On the first iterations, I kept getting bigger grids even though the
grid container fit the specified 500px size (at least on my machine).

In the next commits, I will be implementing the controls to draw on the
grid. These controls will allow the user to change the number of cells,
choose the color to paint with, among other controls I can think of
while I am developing those.