Add ability to draw on mouseover
In this commit, I fixed the drawing issue I was having previously,
namely not being able to draw on mouseover. I implemented a listener
that targets the entire grid container as opposed to each individual
cell. This makes it easier for the container to manage its items.

I also added an erase button that clears up the entire grid and keeps
it at its current size. In order to do this, I had to implement two
variables to keep track of the current grid size and color. The
grid is re-rendered at the current grid size when the Clear button is
pressed and the current color is not affected by this.

This commit also changes the layout of the controls. I may have to
re-implement it because it doesn't look good alongside the grid
container. I am thinking about following the classic Etch-A-Sketch
physical design and add the buttons at the bottom of it to substitute
the two knobs it normally has. This is not certain yet but a possibility.

In the next commit, I am going to add an eraser button. I am still
thinking about how to implement the rainbow option for drawing. This
option picks up a random color on each mouseover event.