Add grid toggle, eraser button, and bug fixes
In this commit, I added more features: a button to toggle the grid
lines on and off as well as an eraser for the user to fix their
mistakes.

Adding these features brought about several bugs, namely clicking
on their buttons and not triggering their intended action sometimes.
It turns out that because I was not keeping track of the state of the
eraser and grid lines, they would take an extra click to perform their
intended action. This happened because as soon as they changed value,
the buttons had to double guess the status of these features. In order
to fix this, I added two variables that can keep track of the state of
both the grid lines and current color.

The toggleGridButton event listener updates currentGridCellBorder
accordingly, and the eraserButon event listener updates both the current
and previous colors according to their values, respectively.

One minor but I am noticing is that when I change the grid size, and
the grid size is bigger than 50 and/or the size is a prime number, it
draws some thick lines. I have to figure out if this is an issue with
CSS Grid or if it's a limitation of the browser when rendering so
many items.

Another bug that needs to be fixed is that when the user clicks on
the eraser and changes the color, they are able to draw with the
new color. However, when they try to use the eraser again, it needs
to be double clicked to be toggled again. I intend to fix this issue
in the next commit.