Fix eraser issue on color change and if active
In the previous commit, I listed an issue with the eraser button --
the user would click on the eraser button, then pick up a new color,
then use the eraser again and not erase the content. While this was
a minor inconvenience, I realized I had to fix this issue as it
could make the user pretty uncomfortable in the long term.

The hotfix I applied consisted in checking for the value of the eraser
button when selecting a new color, and if active, the event listener
for color change would toggle its value and then proceed to update
the current color.

In the next commit, I intend to implement the functionality to change
the background color of the grid without it affecting what is currently
drawn by the user.