Add change background function and bug fixes
On this commit , I added a new feature -- change background color.
What it does is it allows the user to change the background color
of the grid while maintaining what they have currently drawn.

I also improved the clear button functionality. It now clears the
grid and resets back to the default background color. One issue I
found before was that resetting it would maintain the current
background color. I fixed this issue by updating the values for
the background color and current color on the event listener for the
clear button.

I then found another issue when clearing out the grid. It turns out
that if you cleared out the grid, the eraser would stay on, thus
causing the user to draw using the previous background color. I fixed
this by making sure that when Clear is pressed, the default values for
the current and background color are reset and to force the erase
button into the off state.

On the next commit, I am going to work on the rainbow functionality.
This will allow the user to draw with randomly generated colors that
will simulate a rainbow. When this feature is completed, I am going to
work on fixing and improving the UI.