Fix erased cells not being marked as unhovered
On this commit, I fixed an issue where the user would erase any number
of cells, the background color changed, and the erased cells would
retain the previous background color. I did so by including a check
for the erase button every time a cell is hovered. If it is on, it
ensures that if the eraser is on, the cell is marked as unhovered
(hover = false).

This is a bug I discovered when I was making sure multiple changes in
color for both the background and current color would not cause
unexpected results.