Add color working mode and fix Rainbow Mode issues
In this commit, several changes to the behavior of the application were
made. Previously, in order to track in which mode the application was,
I had to rely on currentCellColor and previousCellColor. These
variables would change based on the mode that was currently used. They
were mostly used to implement the functionality of the eraser. Moreover,
the logic behind them relied on tracking whether their button value
was on or off. Doing so introduced bugs when adding extra modes --
the eraser would not perform its task when rainbow mode was also
enabled, even though I had implemented logic to disable rainbow mode
as well.

I started looking for a better way to manage the state of the program.
I still did not want to rely on libraries or complicated things, like
React and Redux at this point. I decided that it would be best to
manage the state of the mode buttons through events. Element.removeListener()
was not removing the events through if statements. The events would
still remain attached to the grid container. This is when I found out
about the AbortController mechanish. This allowed me to remove all
the event listeners attached to the grid container. However, since
all the events are removed, I had to rely on synthetic events to issue
a new event whenever the AbortController removed the painting events.

By using AbortController and synthetic events, we no longer have to
worry about keeping track of the state of each button. This is now
done by adding an .active class to the current button. And when each
button is clicked, it generates its own event to add the corresponding
event listener.

Each mode is now independent, and if we so desired to add more modes,
we could easily do so by including a handler for it.

This took me quite some considerable time to figure out. I could not
think straight and needed to take a break for a few days. Was totally
worth it, as I was able to come up with a cleaner, more efficient
solution.

In the next commits, I intend to change the style of the web page
and make it more aesthetically pleasing. The functionality is now
complete. The website just needs some revamping in the looks department.