Clean up unused logic and variables
In this commit, I updated several comments that were out of date and
removed unused logic that was previously used to enable and keep
track of erase mode. As the behavior of this mode is now decoupled
from that of the other modes, we no longer need to keep track of
when the Erase Mode button is on or off. We also no longer need to
keep track of the previously used color before this mode was enabled,
as the new behavior changes color based on the current background color.

I also furthered the documentation of some functions that were cleaned
up and added more comments that reflect the new changes, namely the
addition of synthetic events along AbortController.

From now on, the next commits should be focused on improving the
user interface of the app (unless a pesky, sneaky bug shows up, of course!).