Fix mode buttons not changing color when clicked
In this commit, I addressed the mode buttons not changing color when
clicked. I also noticed there was an issue with the toggle button.
By default, the grid lines would be on. I wanted to turn the lines
off by default since a lot of people may not be okay with them at
first.

Even though the color mode was the default active mode, it did not
start listening to events until clicked, and it also showed as if
it was currently active. As a result, I decided to only add the
active class to the button modes when clicked and had to change the
event listener for the mode buttons to reflect this behavior.

In the next commits, I intend to perform more style changes to the
page.