Center controls and grid container
In this commit, I finally managed to center the controls panel and
the grid container. The issue was coming from the repeat, auto-fit
hack for the grid columns of the body element. I noticed that if
I hard-coded a specific number of columns greater than three, the
lines would not be displayed correctly.

In order to solve this problem, I had to specify two columns with
fractional units of 1fr on the left and right sides and have the
auto-fit hack handle the sizing required to place the controls and
the grid container. Now these elements are always centered.

As I have been working on this project, I have noticed that the
changes I have been making look great on desktop devices but not
on mobile. As such, in the next few commits, I intend to make
additional modifications tailored towards desktop, and with this,
I will conclude the first version of this project.

I'm hoping that as I become more comfortable with Web development, I
can deliver a version that works well on mobile devices, too.